Added DeleteAll() and HideAll() functions
DeleteAll() will clean up the pipeline, without the need for resetting
the server. This has the advantage of not loosing any python variables,
and evading the ParaView bug which makes it crash when resetting the
server.
HideAll() will simply hide all filters that are in the pipeline.
Everything will still be there, but nothing will be visible. This can
be very useful if a complete grid is added, and one wishes to make it
disappear later.