Major generalization of coordinate system
The coordinate system is now not automatically assumed to be
lon-lat-pressure, but arbitrary x,y,z.
Any axis can be treated as logarithmic. The list ‘logCoord’ denotes
which dimensions are to be considered logarithmic, and the list ‘basis’
denotes the basis for each axis. By default, logCoord=[2] and
basis=[1e3], which corresponds to pressure coordinates. For all linear
coordinates, set logCoord=[]