Example update - ParaView 4.2
Due to a few changes in the filters with ParaView 4.2, the examples had
to be adjusted. These versions will therefore not work in ParaView <
4.2  versions.
In addition, added proper import commands as pv_atmos is now on pypi,
and the files first check if the input files actually exist before
running the code.