Added tools file
tools.py is an addition to pv_atmos, which adds functionality not
needed directly within ParaView. It’s used for pre-processing data and
files, and will receive further functions in the future.
At the moment, it only contains WriteGenericNc(), which allows for
writing a generic netCDF file that can be read with pv_atmos. This is
useful for data that does not initially follow any of the CF
conventions.